AI_Assistant/python/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
pydantic>=2.0.0
python-multipart>=0.0.9